kafka_django/requirements.txt
asgiref==3.4.1
Django==3.2.9
kafka-python==2.0.2
pytz==2021.3
sqlparse==0.4.2
psycopg2==2.9.1
djangorestframework==3.12
django-cors-headers==3.10
